Improved the management of communities and of online resources in shared folders.
Also in a top-level community, the Administrator can now add members on his/her initiative. In the view of communities and projects, the icons associated to the open/close buttons have been swapped: they now match the current state, not the target state. Removed the message "Mentoring is not available for members of this community". The "state" field is now included in the edit form for documents and online resources within folders. Improved the handling of online resources within folders: now you should use the link provided by the "Share" function, in place of the "Share to web" function, as the embed_code of Google docs, sheets and presentations; this allows proper functioning of both the view button in the documents list, and the portlet in the project home.